+++ src/components/testTest/TestTest.module.less
@@ -1,0 +1,4 @@
+@import (reference) "../../references.less";
+
+.root {
+}
